requirements.txt
Django>=2.2.10
django-braces==1.13.0
django-extensions==1.6.7
django-filter==0.13.0
django-filters==0.2.1
django-model-utils==3.0.0
django-rest-swagger==2.0.2
djangorestframework-filters==0.8.0
djangorestframework==3.4.0
simple-crypt==4.1.7
model-mommy>=1.4.0
